Rename route protectors
