+++ src/App.jsx
@@ -8,8 +8,8 @@ import LoginPage from "./pages/LoginPage";
 import SignupPage from "./pages/SignupPage";
 import CreateItemPage from "./pages/CreateItemPage"
 import ProfilePage from "./pages/ProfilePage";
-import IsPrivate from "./components/isPrivate";
-import IsPublic from "./components/isPublic";
+import IsPrivate from "./components/IsPrivate";
+import IsPublic from "./components/IsPublic";
 
 function App() {
 
@@ -23,8 +23,8 @@ function App() {
           <div id="main-content" className="p-4">
             <Routes>
               <Route path="/" element={<Home />} />
-              <Route path="/login" element={<isPublic> <LoginPage /> </isPublic>} />
-              <Route path="/signup" element={<isPublic> <SignupPage /></isPublic> } />
+              <Route path="/login" element={<IsPublic> <LoginPage /> </IsPublic>} />
+              <Route path="/signup" element={<IsPublic> <SignupPage /></IsPublic> } />
               <Route path="/create-item" element={<IsPrivate> <CreateItemPage/> </IsPrivate>} />
